ai-service/requirements.txt
fastapi
uvicorn
sentence-transformers
pydantic
pydantic-settings
pymongo
numpy
python-dotenv
networkx
scikit-learn
matplotlib
pandas
python-multipart
websockets
